fix(models): заменить Float на Numeric для точного хранения денег

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Enum
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, ForeignKey, Enum
 from sqlalchemy.orm import relationship
 from datetime import datetime
 import enum
@@ -17,7 +17,7 @@ class User(Base):
     email = Column(String, unique=True, index=True, nullable=False)
     username = Column(String, unique=True, index=True, nullable=False)
     hashed_password = Column(String, nullable=False)
-    balance = Column(Float, default=100000.0)
+    balance = Column(Numeric(20, 2), default=100000.00)
     created_at = Column(DateTime, default=datetime.utcnow)
 
     portfolio = relationship("Portfolio", back_populates="user", cascade="all, delete-orphan")
@@ -30,7 +30,7 @@ class Stock(Base):
     id = Column(Integer, primary_key=True, index=True)
     symbol = Column(String, unique=True, index=True, nullable=False)
     name = Column(String, nullable=False)
-    current_price = Column(Float, nullable=False, default=0.0)
+    current_price = Column(Numeric(20, 2), nullable=False, default=0.00)
     last_updated = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
 
     portfolio_items = relationship("Portfolio", back_populates="stock")
@@ -43,8 +43,8 @@ class Portfolio(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     stock_id = Column(Integer, ForeignKey("stocks.id"), nullable=False)
-    quantity = Column(Float, nullable=False, default=0.0)
-    average_buy_price = Column(Float, nullable=False, default=0.0)
+    quantity = Column(Numeric(20, 2), nullable=False, default=0.00)
+    average_buy_price = Column(Numeric(20, 2), nullable=False, default=0.00)
 
     user = relationship("User", back_populates="portfolio")
     stock = relationship("Stock", back_populates="portfolio_items")
@@ -57,9 +57,9 @@ class Transaction(Base):
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     stock_id = Column(Integer, ForeignKey("stocks.id"), nullable=False)
     type = Column(Enum(TransactionType), nullable=False)
-    quantity = Column(Float, nullable=False)
-    price_per_share = Column(Float, nullable=False)
-    total_amount = Column(Float, nullable=False)
+    quantity = Column(Numeric(20, 2), nullable=False)
+    price_per_share = Column(Numeric(20, 2), nullable=False)
+    total_amount = Column(Numeric(20, 2), nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
 
     user = relationship("User", back_populates="transactions")
